data-collector/requirements.txt
websocket-client
influxdb-client
